add csrf to users web flow

+++ app/src/foundation/web/routes/users.py
@@ -1,15 +1,14 @@
 from uuid import UUID
 
-from fastapi import Request, Form, Response, APIRouter
-from fastapi import status
+from fastapi import Request, APIRouter, Response
 from sqlalchemy import select
 from starlette.responses import HTMLResponse
 
 from foundation.users.deps import UserServiceDep, UserPaginationDep
 from foundation.users.models import User
-from foundation.users.schemas import Message
 from foundation.users.services import UserValueError, UserNotFoundError
 from foundation.web.deps import CurrentUserDep
+from foundation.web.forms import UserCreateForm, UserEditForm
 from foundation.web.templates import templates
 from foundation.web.utils import flash
 
@@ -30,21 +29,24 @@ async def users(
 
     return templates.TemplateResponse(
         "pages/user_list.html",
-        {
-            "request": request,
-            "page": page,
-        },
+        dict(
+            request=request,
+            page=page,
+        ),
     )
 
 
 @router.get("/users/create")
 async def user_create(
         request: Request,
-        current_user: CurrentUserDep,
 ):
+    form = UserCreateForm(request)
     return templates.TemplateResponse(
         "pages/user_create.html",
-        {"request": request, "current_user": current_user, "user": None},
+        dict(
+            request=request,
+            form=form
+        ),
     )
 
 
@@ -52,44 +54,31 @@ async def user_create(
 async def user_create_post(
         request: Request,
         user_service: UserServiceDep,
-        current_user: CurrentUserDep,
-        full_name: str = Form(),
-        email: str = Form(),
-        password: str = Form(),
-        is_active: bool = Form(False),
-        is_superuser: bool = Form(False),
 ):
+    form = await UserCreateForm.from_formdata(request)
     error = None
-    try:
-        created_user = await user_service.create_user(
-            create_dict={
-                "full_name": full_name,
-                "email": email,
-                "password": password,
-                "is_active": is_active,
-                "is_superuser": is_superuser,
-            }
-        )
-    except UserValueError as e:
-        error = e.args
-        return templates.TemplateResponse(
-            "pages/user_create.html",
-            {
-                "request": request,
-                "error": error,
-                "user": {
-                    "full_name": full_name,
-                    "email": email,
-                    "is_active": is_active,
-                    "is_superuser": is_superuser,
-                },
-            },
-            block_name="content",
-        )
-
+    if await form.validate():
+        try:
+            created_user = await user_service.create_user(
+                create_dict=form.data
+            )
+            return templates.TemplateResponse(
+                "pages/user_view.html",
+                dict(
+                    request=request,
+                    user=created_user
+                ),
+                block_name="content",
+            )
+        except UserValueError as e:
+            error = e.args
     return templates.TemplateResponse(
-        "pages/user_view.html",
-        {"request": request, "error": error, "user": created_user},
+        "pages/user_create.html",
+        dict(
+            request=request,
+            error=error,
+            form=form
+        ),
         block_name="content",
     )
 
@@ -113,14 +102,23 @@ async def user_edit(
         request: Request,
         user_id: UUID,
         user_service: UserServiceDep,
-        current_user: CurrentUserDep,
 ):
     edit_user = await user_service.get_user_by_id(user_id=user_id)
-    return templates.TemplateResponse(
+    form = UserEditForm(request, obj=edit_user)
+
+    # Generate a CSRF token and set it in a cookie
+    token = csrf_token(request)
+    response = templates.TemplateResponse(
         "pages/user_edit.html",
-        {"request": request, "current_user": current_user, "user": edit_user},
+        dict(
+            request=request,
+            user=edit_user,
+            form=form
+        ),
         block_name="content",
     )
+    response.set_cookie("csrf_token", token)  # Set the CSRF token in the cookie
+    return response
 
 
 @router.post("/users/{user_id}")
@@ -129,67 +127,71 @@ async def user_edit_post(
         user_id: UUID,
         user_service: UserServiceDep,
         current_user: CurrentUserDep,
-        full_name: str = Form(),
-        email: str = Form(),
-        is_active: bool = Form(False),
-        is_superuser: bool = Form(False),
 ):
+    form = await UserEditForm.from_formdata(request)
     error = None
-    try:
-        updated_user = await user_service.update_user(
-            user_id=user_id,
-            update_dict={
-                "full_name": full_name,
-                "email": email,
-                "is_active": is_active,
-                "is_superuser": is_superuser,
-            },
-        )
-    except UserValueError as e:
-        error = e.args
-        return templates.TemplateResponse(
-            "pages/user_edit.html",
-            {
-                "request": request,
-                "error": error,
-                "user": {
-                    "full_name": full_name,
-                    "email": email,
-                    "is_active": is_active,
-                    "is_superuser": is_superuser,
-                },
-            },
-            block_name="content",
-        )
+
+    edit_user = await user_service.get_user_by_id(user_id=user_id)
+    if await form.validate():
+        try:
+            updated_user = await user_service.update_user(
+                user_id=user_id,
+                update_dict=form.data,
+            )
+            return templates.TemplateResponse(
+                "pages/user_view.html",
+                dict(
+                    request=request,
+                    user=updated_user
+                ),
+                block_name="content",
+            )
+        except UserValueError as e:
+            error = e.args
 
     return templates.TemplateResponse(
-        "pages/user_view.html",
-        {"request": request, "error": error, "user": updated_user},
+        "pages/user_edit.html",
+        dict(
+            request=request,
+            user=edit_user,
+            error=error,
+            form=form
+        ),
         block_name="content",
     )
+
+
+from fastapi import Request, Header, HTTPException
+from fastapi import status
+from starlette_wtf import csrf_token
 
 
 @router.delete("/users/{user_id}")
 async def delete_user(
         request: Request,
-        user_service: UserServiceDep,
         user_id: UUID,
+        user_service: UserServiceDep,
         current_user: CurrentUserDep,
+        x_csrftoken: str = Header(None)
 ):
     """
-    Delete a user.
-    If the current_user is a non-super user, they can only delete their own user.
-    Superusers can delete any other user.
-    Superusers can not delete their own user.
+    Delete a user. Only admins can delete users.
     """
     if not current_user.is_superuser:
-        return Message(message="Only admins can delete users.")
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Only admins can delete users.")
+
+    # Retrieve the CSRF token from the cookie
+    csrf_token_in_cookie = request.cookies.get("csrf_token")
+
+    # Validate the CSRF token directly from the header
+    if not x_csrftoken or x_csrftoken != csrf_token_in_cookie:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Invalid CSRF token.")
 
     try:
         user = await user_service.get_user_by_id(user_id=user_id)
         await user_service.delete_user(user_id=user_id)
     except UserNotFoundError as e:
-        return Message(message=e.args[0])
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=e.args[0])
 
     flash(request, f"User {user.full_name} was deleted")
     response = Response(status_code=status.HTTP_307_TEMPORARY_REDIRECT)
